requirements-dev.txt
amqp==5.0.6
anyio==3.1.0
arabic-reshaper==2.1.3
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.17.94
botocore==1.20.94
celery==5.1.0
certifi==2021.5.30
click==8.0.1
click-plugins==1.1.1
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap-form==3.4
django-bootstrap4==3.0.1
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-debug-permissions==1.0.0
django-storages==1.11.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
future==0.18.2
h11==0.12.0
html5lib==1.1
httpcore==0.13.5
hyperlink==21.0.0
incremental==21.3.0
jmespath==0.10.0
kombu==5.1.0
Pillow==8.2.0
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
redis==3.5.3
reportlab==3.5.67
requests==2.25.1
rfc3986==1.5.0
s3transfer==0.4.2
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
